% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLDataCalc_aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate an LPJmLDataCalc object}
\usage{
aggregate(
  lpjml_calc,
  space = "cow_regions",
  method = "sum",
  support_of_area_dens = "terr_area"
)
}
\arguments{
\item{lpjml_calc}{\code{\link{LPJmLDataCalc}} object to be aggregated}

\item{space}{The spatial aggregation units. Can be either a string
indicating regions of the earth surface or an \code{\link{LPJmLRegionData}} object.
Currently the following strings are available:
\itemize{
\item \code{cow_regions}: The regions defined by the Correlates of War project
\item \code{global}: A dynamically created region that fully contains all cells
of the grid.
}}

\item{method}{String that specifies spatial method of aggregation.
Current only the following options are available:
\itemize{
\item \code{sum}: The values of all cells belonging to each region are summed up.
If a cell belongs to a region only partially, we assume
that the quantity is distributed uniformly over the cell area and
multiply the value by the fraction of the cell that is part of the region
before summing up.
\item \code{integral}: Given an area density, the integral over each region is
calculated. First, the area density is converted to a total per cell by
multiplying the value by the supporting cell areas.
The integral is then calculated
by summing up the values of all cells belonging to each region, as described
for \code{sum}. This option expects that the stored quantity is an
area density, i.e. \code{m^2} is par of the denominator of the unit.
\item \code{area_mean}: Calculates the mean value over the supporting areas of each
region. This option expects that the stored quantity is an
area density, i.e. \code{m^2} is par of the denominator of the unit.
}}

\item{support_of_area_dens}{Specifies the areas of each cell on which
the area density is supported. See \code{\link{LPJmLDataCalc}} method
area_dens2cell_values.}
}
\value{
An aggregated \code{\link{LPJmLDataCalc}} object.
}
\description{
Function to aggregate an \code{\link{LPJmLDataCalc}} object in space.
Different aggregation methods can be used.
}
