% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLDataCalc.R
\name{as_LPJmLDataCalc}
\alias{as_LPJmLDataCalc}
\title{Coerce an LPJmLData object into an LPJmLDataCalc object}
\usage{
as_LPJmLDataCalc(obj)
}
\arguments{
\item{obj}{LPJmLData object.}
}
\value{
An LPJmLDataCalc object.
}
\description{
Function to coerce (convert) an \code{\link{LPJmLData}} object into an
LPJmLDataCalc object with extended functionality.
}
