% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{Benchmark one or several LPJmL runs}
\usage{
benchmark(bsl_dir, ut_dirs, settings, file_type = ".bin.json")
}
\arguments{
\item{bsl_dir}{Path to directory containing the baseline run.}

\item{ut_dirs}{List of paths to directories containing the under test run
results.}

\item{settings}{List that defines for each output which metrics to use.
The list has to have the following structure:
\itemize{
\item \code{var1}: List of metric classes to use for variable \code{var1}
\item \code{var2}: List of metric classes to use for variable \code{var2}
\item ...
}}

\item{file_type}{File type of the raw data to be read in (e.g. ".bin.json")}
}
\value{
List of metric objects which store the results of the evaluations.
}
\description{
Function to benchmark one or several under test LPJmL runs
against a baseline run.
}
