% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLRegionData.R
\name{build_global_region}
\alias{build_global_region}
\title{Construct global region object that fully contains all cells given in a
grid.}
\usage{
build_global_region(grid)
}
\arguments{
\item{grid}{An LPJmLGridData object containing the grid.}
}
\value{
An LPJmLRegionData object containing the global region.
}
\description{
Construct global region object that fully contains all cells given in a
grid.
}
\seealso{
\code{\link{LPJmLRegionData}}
}
