% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLRegionData.R
\name{construct_lpjml_region_global}
\alias{construct_lpjml_region_global}
\title{Construct global region object that fully contains all cells given in a
grid.}
\usage{
construct_lpjml_region_global(grid)
}
\arguments{
\item{grid}{An LPJmLGridData object containing the grid.}
}
\value{
An LPJmLRegionData object containing the global region.
}
\description{
Construct global region object that fully contains all cells given in a
grid.
}
\seealso{
\code{\link{LPJmLRegionData}}
}
