% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ilamb_report.R
\name{create_ilamb_report}
\alias{create_ilamb_report}
\title{Create ILAMB Report}
\usage{
create_ilamb_report(
  baseline_dir = baseline_dir,
  under_test_dirs = under_test_dirs,
  sim_table = NULL,
  output_file = NULL,
  eval_vars = c("mgpp", "mevap", "mtransp", "mrh", "mnpp", "vegc"),
  ilamb_run_script = file.path(ilamb_dir, "ilamb_run_cmd.sh")
)
}
\description{
Create ILAMB Report
}
\keyword{internal}
