% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{create_literature_pdf}
\alias{create_literature_pdf}
\title{Function to create a pdf with a table with literature values}
\usage{
create_literature_pdf(output_file = "literature_values.pdf", ...)
}
\arguments{
\item{output_file}{filename of the output pdf, can include directory}

\item{...}{additional parameters passed to rmarkdown::render}
}
\description{
Function to create a pdf with a table with literature values
}
