% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{create_pdf_report}
\alias{create_pdf_report}
\title{Generate a pdf report from a benchmarkResult object.}
\usage{
create_pdf_report(data, output_file = "benchmark.pdf", ...)
}
\arguments{
\item{data}{benchmarkResult object created by the \code{\link{benchmark}}
function}

\item{output_file}{file of the output pdf, including filename and directory
\code{\link[rmarkdown]{render}}}

\item{...}{additional arguments passed to \code{\link[rmarkdown]{render}}}
}
\description{
Generate a pdf report from a benchmarkResult object.
}
\details{
Each metric has its own section in the report. The content of the section
is generated by the plot and plot_arrange function of the metric.
The metric results are displayed in the same order as they were specified in
the benchmark settings.
}
\examples{
\dontrun{
  create_pdf_report(BM_data, "myBenchmark.pdf")
}



}
