% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_settings.R
\docType{data}
\name{default_settings}
\alias{default_settings}
\title{Default settings for the Benchmarking}
\format{
An object of class \code{list} of length 29.
}
\usage{
default_settings
}
\description{
Default settings for the Benchmarking
}
\keyword{datasets}
