% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLDataCalc.R
\name{-.LPJmLDataCalc}
\alias{-.LPJmLDataCalc}
\title{Subtraction of two LPJmLDataCalc objects}
\usage{
\method{-}{LPJmLDataCalc}(o1, o2)
}
\arguments{
\item{o1}{An \code{LPJmLDataCalc} object.}

\item{o2}{An \code{LPJmLDataCalc} object.}
}
\value{
An \code{LPJmLDataCalc} object.
}
\description{
Subtract an LPJmLDataCalc object from another LPJmLDataCalc object
}
