% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metric.R
\name{GlobalSumTimeAvg}
\alias{GlobalSumTimeAvg}
\title{GlobalSumTimeAvg}
\description{
GlobalSumTimeAvg metric.
}
\section{Super class}{
\code{\link[lpjmlstats:Metric]{lpjmlstats::Metric}} -> \code{GlobalSumTimeAvg}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GlobalSumTimeAvg-summarize}{\code{GlobalSumTimeAvg$summarize()}}
\item \href{#method-GlobalSumTimeAvg-compare}{\code{GlobalSumTimeAvg$compare()}}
\item \href{#method-GlobalSumTimeAvg-clone}{\code{GlobalSumTimeAvg$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="add_comparison"><a href='../../lpjmlstats/html/Metric.html#method-Metric-add_comparison'><code>lpjmlstats::Metric$add_comparison()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="capture_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-capture_summary'><code>lpjmlstats::Metric$capture_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="store_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-store_summary'><code>lpjmlstats::Metric$store_summary()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalSumTimeAvg-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalSumTimeAvg-summarize}{}}}
\subsection{Method \code{summarize()}}{
Compute a global area weighted sum of a variable and then average
all timesteps of the simulation period.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalSumTimeAvg$summarize(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Raw data to be summarized}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalSumTimeAvg-compare"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalSumTimeAvg-compare}{}}}
\subsection{Method \code{compare()}}{
Compare the baseline summary with the under test summaries by
subtracting the baseline from the under test.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalSumTimeAvg$compare(eval_grp)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{eval_grp}}{Evaluation group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalSumTimeAvg-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalSumTimeAvg-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalSumTimeAvg$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
