% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metric.R
\name{GlobalSumTimeSer}
\alias{GlobalSumTimeSer}
\title{GlobalSumTimeSer}
\description{
GlobalSumTimeSer metric.
}
\section{Super class}{
\code{\link[lpjmlstats:Metric]{lpjmlstats::Metric}} -> \code{GlobalSumTimeSer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GlobalSumTimeSer-summarize}{\code{GlobalSumTimeSer$summarize()}}
\item \href{#method-GlobalSumTimeSer-clone}{\code{GlobalSumTimeSer$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="add_comparison"><a href='../../lpjmlstats/html/Metric.html#method-Metric-add_comparison'><code>lpjmlstats::Metric$add_comparison()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="capture_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-capture_summary'><code>lpjmlstats::Metric$capture_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="compare"><a href='../../lpjmlstats/html/Metric.html#method-Metric-compare'><code>lpjmlstats::Metric$compare()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="store_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-store_summary'><code>lpjmlstats::Metric$store_summary()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalSumTimeSer-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalSumTimeSer-summarize}{}}}
\subsection{Method \code{summarize()}}{
Compute a global area weighted sum of a variable
for each timestep.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalSumTimeSer$summarize(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Raw data to be summarized}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalSumTimeSer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalSumTimeSer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalSumTimeSer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
