% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metric_subclasses.R
\name{GlobCellSumAnnAvgTimeseries}
\alias{GlobCellSumAnnAvgTimeseries}
\title{GlobCellSumAnnAvgTimeseries}
\description{
GlobCellSumAnnAvgTimeseries metric
}
\section{Super classes}{
\code{\link[lpjmlstats:Metric]{lpjmlstats::Metric}} -> \code{\link[lpjmlstats:GlobSumTimeseries]{lpjmlstats::GlobSumTimeseries}} -> \code{\link[lpjmlstats:GlobSumAnnAvgTimeseries]{lpjmlstats::GlobSumAnnAvgTimeseries}} -> \code{GlobCellSumAnnAvgTimeseries}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GlobCellSumAnnAvgTimeseries-summarize}{\code{GlobCellSumAnnAvgTimeseries$summarize()}}
\item \href{#method-GlobCellSumAnnAvgTimeseries-clone}{\code{GlobCellSumAnnAvgTimeseries$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="add_compare_meta"><a href='../../lpjmlstats/html/Metric.html#method-Metric-add_compare_meta'><code>lpjmlstats::Metric$add_compare_meta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="add_comparisons"><a href='../../lpjmlstats/html/Metric.html#method-Metric-add_comparisons'><code>lpjmlstats::Metric$add_comparisons()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="capture_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-capture_summary'><code>lpjmlstats::Metric$capture_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="compare"><a href='../../lpjmlstats/html/Metric.html#method-Metric-compare'><code>lpjmlstats::Metric$compare()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="generate_report_content"><a href='../../lpjmlstats/html/Metric.html#method-Metric-generate_report_content'><code>lpjmlstats::Metric$generate_report_content()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="print_metric_description"><a href='../../lpjmlstats/html/Metric.html#method-Metric-print_metric_description'><code>lpjmlstats::Metric$print_metric_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="print_metric_header"><a href='../../lpjmlstats/html/Metric.html#method-Metric-print_metric_header'><code>lpjmlstats::Metric$print_metric_header()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="print_year_subset"><a href='../../lpjmlstats/html/Metric.html#method-Metric-print_year_subset'><code>lpjmlstats::Metric$print_year_subset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="store_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-store_summary'><code>lpjmlstats::Metric$store_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="transform_lpjml_calcs"><a href='../../lpjmlstats/html/Metric.html#method-Metric-transform_lpjml_calcs'><code>lpjmlstats::Metric$transform_lpjml_calcs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="GlobSumTimeseries" data-id="arrange_plots"><a href='../../lpjmlstats/html/GlobSumTimeseries.html#method-GlobSumTimeseries-arrange_plots'><code>lpjmlstats::GlobSumTimeseries$arrange_plots()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="GlobSumTimeseries" data-id="plot"><a href='../../lpjmlstats/html/GlobSumTimeseries.html#method-GlobSumTimeseries-plot'><code>lpjmlstats::GlobSumTimeseries$plot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobCellSumAnnAvgTimeseries-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-GlobCellSumAnnAvgTimeseries-summarize}{}}}
\subsection{Method \code{summarize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobCellSumAnnAvgTimeseries$summarize(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobCellSumAnnAvgTimeseries-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GlobCellSumAnnAvgTimeseries-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobCellSumAnnAvgTimeseries$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
