% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metric_subclasses.R
\name{GlobCellSumTimeAvgTable}
\alias{GlobCellSumTimeAvgTable}
\title{GlobCellSumTimeAvgTable}
\description{
GlobCellSumTimeAvgTable metric
}
\section{Super classes}{
\code{\link[lpjmlstats:Metric]{lpjmlstats::Metric}} -> \code{\link[lpjmlstats:GlobSumTimeAvgTable]{lpjmlstats::GlobSumTimeAvgTable}} -> \code{GlobCellSumTimeAvgTable}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GlobCellSumTimeAvgTable-summarize}{\code{GlobCellSumTimeAvgTable$summarize()}}
\item \href{#method-GlobCellSumTimeAvgTable-clone}{\code{GlobCellSumTimeAvgTable$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="add_compare_meta"><a href='../../lpjmlstats/html/Metric.html#method-Metric-add_compare_meta'><code>lpjmlstats::Metric$add_compare_meta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="add_comparisons"><a href='../../lpjmlstats/html/Metric.html#method-Metric-add_comparisons'><code>lpjmlstats::Metric$add_comparisons()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="capture_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-capture_summary'><code>lpjmlstats::Metric$capture_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="generate_report_content"><a href='../../lpjmlstats/html/Metric.html#method-Metric-generate_report_content'><code>lpjmlstats::Metric$generate_report_content()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="print_metric_description"><a href='../../lpjmlstats/html/Metric.html#method-Metric-print_metric_description'><code>lpjmlstats::Metric$print_metric_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="print_metric_header"><a href='../../lpjmlstats/html/Metric.html#method-Metric-print_metric_header'><code>lpjmlstats::Metric$print_metric_header()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="print_year_subset"><a href='../../lpjmlstats/html/Metric.html#method-Metric-print_year_subset'><code>lpjmlstats::Metric$print_year_subset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="store_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-store_summary'><code>lpjmlstats::Metric$store_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="transform_lpjml_calcs"><a href='../../lpjmlstats/html/Metric.html#method-Metric-transform_lpjml_calcs'><code>lpjmlstats::Metric$transform_lpjml_calcs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="GlobSumTimeAvgTable" data-id="arrange_plots"><a href='../../lpjmlstats/html/GlobSumTimeAvgTable.html#method-GlobSumTimeAvgTable-arrange_plots'><code>lpjmlstats::GlobSumTimeAvgTable$arrange_plots()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="GlobSumTimeAvgTable" data-id="compare"><a href='../../lpjmlstats/html/GlobSumTimeAvgTable.html#method-GlobSumTimeAvgTable-compare'><code>lpjmlstats::GlobSumTimeAvgTable$compare()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="GlobSumTimeAvgTable" data-id="plot"><a href='../../lpjmlstats/html/GlobSumTimeAvgTable.html#method-GlobSumTimeAvgTable-plot'><code>lpjmlstats::GlobSumTimeAvgTable$plot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobCellSumTimeAvgTable-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-GlobCellSumTimeAvgTable-summarize}{}}}
\subsection{Method \code{summarize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobCellSumTimeAvgTable$summarize(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobCellSumTimeAvgTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GlobCellSumTimeAvgTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobCellSumTimeAvgTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
