% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metric_subclasses.R
\name{GlobSumAnnAvgTimeseries}
\alias{GlobSumAnnAvgTimeseries}
\title{GlobSumAnnAvgTimeseries}
\description{
GlobSumAnnAvgTimeseries metric
}
\section{Super classes}{
\code{\link[lpjmlstats:Metric]{lpjmlstats::Metric}} -> \code{\link[lpjmlstats:GlobSumTimeseries]{lpjmlstats::GlobSumTimeseries}} -> \code{GlobSumAnnAvgTimeseries}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{Section header used in the report}

\item{\code{description}}{Description used in the report}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GlobSumAnnAvgTimeseries-summarize}{\code{GlobSumAnnAvgTimeseries$summarize()}}
\item \href{#method-GlobSumAnnAvgTimeseries-clone}{\code{GlobSumAnnAvgTimeseries$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="add_comparison"><a href='../../lpjmlstats/html/Metric.html#method-Metric-add_comparison'><code>lpjmlstats::Metric$add_comparison()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="apply_to_all_lpjml_calcs"><a href='../../lpjmlstats/html/Metric.html#method-Metric-apply_to_all_lpjml_calcs'><code>lpjmlstats::Metric$apply_to_all_lpjml_calcs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="capture_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-capture_summary'><code>lpjmlstats::Metric$capture_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="compare"><a href='../../lpjmlstats/html/Metric.html#method-Metric-compare'><code>lpjmlstats::Metric$compare()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="generate_report_content"><a href='../../lpjmlstats/html/Metric.html#method-Metric-generate_report_content'><code>lpjmlstats::Metric$generate_report_content()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="print_metric_description"><a href='../../lpjmlstats/html/Metric.html#method-Metric-print_metric_description'><code>lpjmlstats::Metric$print_metric_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="print_metric_header"><a href='../../lpjmlstats/html/Metric.html#method-Metric-print_metric_header'><code>lpjmlstats::Metric$print_metric_header()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="store_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-store_summary'><code>lpjmlstats::Metric$store_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="GlobSumTimeseries" data-id="arrange_plots"><a href='../../lpjmlstats/html/GlobSumTimeseries.html#method-GlobSumTimeseries-arrange_plots'><code>lpjmlstats::GlobSumTimeseries$arrange_plots()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="GlobSumTimeseries" data-id="plot"><a href='../../lpjmlstats/html/GlobSumTimeseries.html#method-GlobSumTimeseries-plot'><code>lpjmlstats::GlobSumTimeseries$plot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobSumAnnAvgTimeseries-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-GlobSumAnnAvgTimeseries-summarize}{}}}
\subsection{Method \code{summarize()}}{
Take the mean for each year and then the global weighted sum
over the cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobSumAnnAvgTimeseries$summarize(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{LpjmlDataCalc object to be summarized}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A summarized \link{LPJmLDataCalc} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobSumAnnAvgTimeseries-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GlobSumAnnAvgTimeseries-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobSumAnnAvgTimeseries$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
