% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metric_subclasses.R
\name{GlobSumTimeAvgTablePFT_harvest}
\alias{GlobSumTimeAvgTablePFT_harvest}
\title{GlobSumTimeAvgTablePFT_harvest}
\description{
GlobSumTimeAvgTablePFT_harvest metric
}
\section{Super classes}{
\code{\link[lpjmlstats:Metric]{lpjmlstats::Metric}} -> \code{\link[lpjmlstats:GlobSumTimeAvgTable]{lpjmlstats::GlobSumTimeAvgTable}} -> \code{GlobSumTimeAvgTablePFT_harvest}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{Section header used in the report}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GlobSumTimeAvgTablePFT_harvest-summarize}{\code{GlobSumTimeAvgTablePFT_harvest$summarize()}}
\item \href{#method-GlobSumTimeAvgTablePFT_harvest-new}{\code{GlobSumTimeAvgTablePFT_harvest$new()}}
\item \href{#method-GlobSumTimeAvgTablePFT_harvest-clone}{\code{GlobSumTimeAvgTablePFT_harvest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="add_compare_meta"><a href='../../lpjmlstats/html/Metric.html#method-Metric-add_compare_meta'><code>lpjmlstats::Metric$add_compare_meta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="add_comparisons"><a href='../../lpjmlstats/html/Metric.html#method-Metric-add_comparisons'><code>lpjmlstats::Metric$add_comparisons()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="capture_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-capture_summary'><code>lpjmlstats::Metric$capture_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="generate_report_content"><a href='../../lpjmlstats/html/Metric.html#method-Metric-generate_report_content'><code>lpjmlstats::Metric$generate_report_content()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="print_metric_description"><a href='../../lpjmlstats/html/Metric.html#method-Metric-print_metric_description'><code>lpjmlstats::Metric$print_metric_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="print_metric_header"><a href='../../lpjmlstats/html/Metric.html#method-Metric-print_metric_header'><code>lpjmlstats::Metric$print_metric_header()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="print_year_subset"><a href='../../lpjmlstats/html/Metric.html#method-Metric-print_year_subset'><code>lpjmlstats::Metric$print_year_subset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="store_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-store_summary'><code>lpjmlstats::Metric$store_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="transform_lpjml_calcs"><a href='../../lpjmlstats/html/Metric.html#method-Metric-transform_lpjml_calcs'><code>lpjmlstats::Metric$transform_lpjml_calcs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="GlobSumTimeAvgTable" data-id="arrange_plots"><a href='../../lpjmlstats/html/GlobSumTimeAvgTable.html#method-GlobSumTimeAvgTable-arrange_plots'><code>lpjmlstats::GlobSumTimeAvgTable$arrange_plots()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="GlobSumTimeAvgTable" data-id="compare"><a href='../../lpjmlstats/html/GlobSumTimeAvgTable.html#method-GlobSumTimeAvgTable-compare'><code>lpjmlstats::GlobSumTimeAvgTable$compare()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="GlobSumTimeAvgTable" data-id="plot"><a href='../../lpjmlstats/html/GlobSumTimeAvgTable.html#method-GlobSumTimeAvgTable-plot'><code>lpjmlstats::GlobSumTimeAvgTable$plot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobSumTimeAvgTablePFT_harvest-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-GlobSumTimeAvgTablePFT_harvest-summarize}{}}}
\subsection{Method \code{summarize()}}{
Weigh by cft_frac and then do the same as GlobSumTimeAvgTable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobSumTimeAvgTablePFT_harvest$summarize(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{LPJmLDataCalc object to be summarized}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobSumTimeAvgTablePFT_harvest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GlobSumTimeAvgTablePFT_harvest-new}{}}}
\subsection{Method \code{new()}}{
initialize with an extended description
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobSumTimeAvgTablePFT_harvest$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobSumTimeAvgTablePFT_harvest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GlobSumTimeAvgTablePFT_harvest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobSumTimeAvgTablePFT_harvest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
