% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLDataCalc.R
\name{LPJmLDataCalc}
\alias{LPJmLDataCalc}
\title{LPJmLDataCalc}
\description{
An extended LPJmLData class that enables calculations and statistics.
}
\section{Super class}{
\code{\link[lpjmlkit:LPJmLData]{lpjmlkit::LPJmLData}} -> \code{LPJmLDataCalc}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LPJmLDataCalc-new}{\code{LPJmLDataCalc$new()}}
\item \href{#method-LPJmLDataCalc-add}{\code{LPJmLDataCalc$add()}}
\item \href{#method-LPJmLDataCalc-subtract}{\code{LPJmLDataCalc$subtract()}}
\item \href{#method-LPJmLDataCalc-multiply}{\code{LPJmLDataCalc$multiply()}}
\item \href{#method-LPJmLDataCalc-divide}{\code{LPJmLDataCalc$divide()}}
\item \href{#method-LPJmLDataCalc-clone}{\code{LPJmLDataCalc$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id=".__set_data__"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-.__set_data__'><code>lpjmlkit::LPJmLData$.__set_data__()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id=".__set_grid__"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-.__set_grid__'><code>lpjmlkit::LPJmLData$.__set_grid__()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="add_grid"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-add_grid'><code>lpjmlkit::LPJmLData$add_grid()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="as_array"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-as_array'><code>lpjmlkit::LPJmLData$as_array()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="as_raster"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-as_raster'><code>lpjmlkit::LPJmLData$as_raster()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="as_terra"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-as_terra'><code>lpjmlkit::LPJmLData$as_terra()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="as_tibble"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-as_tibble'><code>lpjmlkit::LPJmLData$as_tibble()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="dim"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-dim'><code>lpjmlkit::LPJmLData$dim()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="dimnames"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-dimnames'><code>lpjmlkit::LPJmLData$dimnames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="length"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-length'><code>lpjmlkit::LPJmLData$length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="plot"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-plot'><code>lpjmlkit::LPJmLData$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="print"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-print'><code>lpjmlkit::LPJmLData$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="subset"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-subset'><code>lpjmlkit::LPJmLData$subset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="summary"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-summary'><code>lpjmlkit::LPJmLData$summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="transform"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-transform'><code>lpjmlkit::LPJmLData$transform()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-new"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-new}{}}}
\subsection{Method \code{new()}}{
!Internal method only to be used for package development!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$new(lpjml_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lpjml_data}}{an LPJmLData object.
Addition of two LPJmLDataCalc objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-add"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-add}{}}}
\subsection{Method \code{add()}}{
!Internal method only to be used for package development!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$add(lpjml_calc_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lpjml_calc_obj}}{An \code{LPJmLData} object.
Subtraction of two LPJmLDataCalc objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-subtract"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-subtract}{}}}
\subsection{Method \code{subtract()}}{
!Internal method only to be used for package development!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$subtract(lpjml_calc_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lpjml_calc_obj}}{An \code{LPJmLData} object.
Multiplication of two LPJmLDataCalc objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-multiply"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-multiply}{}}}
\subsection{Method \code{multiply()}}{
!Internal method only to be used for package development!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$multiply(lpjml_calc_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lpjml_calc_obj}}{An \code{LPJmLData} object.
Division of two LPJmLDataCalc objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-divide"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-divide}{}}}
\subsection{Method \code{divide()}}{
!Internal method only to be used for package development!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$divide(lpjml_calc_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lpjml_calc_obj}}{An \code{LPJmLData} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
