% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLRegionData.R
\name{LPJmLRegionData}
\alias{LPJmLRegionData}
\title{LPJmLRegionData}
\description{
A class that represents one or several regions in LPJmL.
Based on an LPJmL grid, a region is defined as set of grid cells together
with fractions. The fractions indicate the share of each grid cell that
is part of the region. (e.g. 1 = the cell belongs completely to the region,
0 = the cell does not belong to the region at all).\
The underlying data structure is a sparse matrix, where the rows represent
the regions, the columns represent the grid cells and the values represent
the fractions (cells not belonging to a region do not take memory as only
nonzero entries are stored in a sparse matrix).

Create a new \code{LPJmLRegionData} object; only used internally.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{region_matrix}}{object stores the region data as a sparse matrix.}

\item{\code{grid}}{\code{LPJmLGridData} object containing the underlying grid.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LPJmLRegionData-new}{\code{LPJmLRegionData$new()}}
\item \href{#method-LPJmLRegionData-get_ncells_per_region}{\code{LPJmLRegionData$get_ncells_per_region()}}
\item \href{#method-LPJmLRegionData-clone}{\code{LPJmLRegionData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLRegionData-new"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLRegionData-new}{}}}
\subsection{Method \code{new()}}{
!Internal method only to be used by the package itself!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLRegionData$new(grid, region_matrix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{grid}}{\code{LPJmLGridData} object containing the underlying grid.}

\item{\code{region_matrix}}{object stores the region data as a sparse matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLRegionData-get_ncells_per_region"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLRegionData-get_ncells_per_region}{}}}
\subsection{Method \code{get_ncells_per_region()}}{
Get number of cells per region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLRegionData$get_ncells_per_region()}\if{html}{\out{</div>}}
}

\subsection{Details}{
For partially belonging cells the fraction
of the cell that belongs to the region is counted.
}

\subsection{Returns}{
A vector of length nrow(region_matrix) containing the number of
cells per region.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLRegionData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLRegionData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLRegionData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
