% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpjmlstats-package.R
\docType{package}
\name{lpjmlstats-package}
\alias{lpjmlstats}
\alias{lpjmlstats-package}
\title{lpjmlstats: Statistical tools for LPJmL data analysis}
\description{
This package provides statistical tools for LPJmL data analysis to be used for benchmarking LPJmL outputs.
}
\author{
\strong{Maintainer}: David Hötten \email{davidho@pik-potsdam.de}

Authors:
\itemize{
  \item Jannes Breier \email{jannes.breier@pik-potsdam.de}
}

}
\keyword{internal}
