% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metric.R
\name{Metric}
\alias{Metric}
\title{Metric}
\description{
A Metric is a class defines a procedure to evaluate and compare
one or several under test outputs against a baseline output.
It is also a container for the results of all evaluations, that have been
performed with this metric.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{evalList}}{List of evaluation groups. Each evaluation group contains the summaries
and the comparisons for one variable.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Metric-capture_summary}{\code{Metric$capture_summary()}}
\item \href{#method-Metric-store_summary}{\code{Metric$store_summary()}}
\item \href{#method-Metric-add_comparison}{\code{Metric$add_comparison()}}
\item \href{#method-Metric-summarize}{\code{Metric$summarize()}}
\item \href{#method-Metric-compare}{\code{Metric$compare()}}
\item \href{#method-Metric-clone}{\code{Metric$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-capture_summary"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-capture_summary}{}}}
\subsection{Method \code{capture_summary()}}{
Compute and store an individual summary of the raw data presented.
Uses the internal summary function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$capture_summary(data, var, type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Raw data to be summarized}

\item{\code{var}}{Variable name}

\item{\code{type}}{Type of data ("bsl" or "ut")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-store_summary"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-store_summary}{}}}
\subsection{Method \code{store_summary()}}{
Store a summary in the metric object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$store_summary(summary, var, type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summary}}{Summary to be stored}

\item{\code{var}}{Variable name}

\item{\code{type}}{Type of data ("bsl" or "ut")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-add_comparison"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-add_comparison}{}}}
\subsection{Method \code{add_comparison()}}{
Compare the baseline summary with the under test summaries and stores
the results in the metric object.
Uses the internal comparison function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$add_comparison(var)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{Variable name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-summarize}{}}}
\subsection{Method \code{summarize()}}{
Pipeline to summarize the raw data.
Will be overwritten by the individual metric classes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$summarize(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Raw data to be summarized}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-compare"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-compare}{}}}
\subsection{Method \code{compare()}}{
Pipeline to compare the baseline summary with the under test summaries.
Will be overwritten by the individual metric classes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$compare(eval_grp)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{eval_grp}}{Evaluation group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
