% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLDataCalc.R
\name{plot.LPJmLDataCalc}
\alias{plot.LPJmLDataCalc}
\title{Plot an LPJmLDataCalc object}
\usage{
\method{plot}{LPJmLDataCalc}(x, ...)
}
\arguments{
\item{x}{LPJmLDataCalc object. In case of non-aggregated data
\link[lpjmlkit]{plot.LPJmLData}
is directly called. In case of aggregated data the
value for each region is assigned to all pixels that belong to the region.
If a pixel belong to a region only partially, the value is multiplied by the
fraction of that pixel belonging to the region. If a pixel belongs to
multiple regions, the sum of all respective region values
(multiplied by the fractions) is taken.
The pixel values are then again plotted with \link[lpjmlkit]{plot.LPJmLData}.}

\item{...}{Arguments passed to \link[lpjmlkit]{LPJmLData} plot method.}
}
\description{
The function acts a wrapper of \link[lpjmlkit]{plot.LPJmLData} from lpjmlkit,
but allows for plotting data in more formats.
}
