% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLRegionData.R
\name{read_cow_regions}
\alias{read_cow_regions}
\title{Read or create the cow regions as an LPJmLRegionData object}
\usage{
read_cow_regions()
}
\value{
An LPJmLRegionData object containing the cow regions.
}
\description{
The COW = countries of the world data contains global country borders.
}
\seealso{
\code{\link{LPJmLRegionData}}
}
