% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLRegionData.R
\name{read_def_grid}
\alias{read_def_grid}
\title{Read default grid}
\usage{
read_def_grid()
}
\value{
An LPJmLGridData object containing the default grid.
}
\description{
The default grid is the standard global grid used in LPJmL.
}
