% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLDataCalc.R
\name{read_io}
\alias{read_io}
\title{Read in LPJmL input and output files as LPJmLDataCalc}
\usage{
read_io(..., output_type = "LPJmLDataCalc")
}
\arguments{
\item{...}{Parameters that are passed to \link[lpjmlkit]{read_io}}

\item{output_type}{Can be either \code{LPJmLDataCalc}or \code{LPJmLData}.}
}
\value{
An \code{\link{LPJmLDataCalc}} object
}
\description{
The function acts a wrapper of \link[lpjmlkit]{read_io} from lpjmlkit,
but outputs an \code{\link{LPJmLDataCalc}} object.
}
