% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLRegionData.R
\name{read_lpjml_grid_default}
\alias{read_lpjml_grid_default}
\title{Read default grid}
\usage{
read_lpjml_grid_default()
}
\value{
An LPJmLGridData object containing the default grid.
}
\description{
The default grid is the standard global grid used in LPJmL.
}
\seealso{
\code{\link{LPJmLGridData}}
}
