% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLRegionData.R
\name{read_lpjml_region_cow}
\alias{read_lpjml_region_cow}
\title{Read or create the cow regions as an LPJmLRegionData object}
\usage{
read_lpjml_region_cow()
}
\value{
An LPJmLRegionData object containing the cow regions.
}
\description{
The correlates of war (COW) data contains global country borders.
In the version used here each gridcell is assigned to
one or more countries.
}
\seealso{
\code{\link{LPJmLRegionData}}
}
