% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLRegionData.R
\name{read_lpjml_region_global_land}
\alias{read_lpjml_region_global_land}
\title{Read or create the global land as an LPJmLRegionData object}
\usage{
read_lpjml_region_global_land()
}
\value{
An LPJmLRegionData object containing the global land.
}
\description{
Global land includes inland water bodies but excludes ocean fraction of cells
along coastlines.
}
\seealso{
\code{\link{LPJmLRegionData}}
}
