% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_settings.R
\name{set_lpjmlstats_settings}
\alias{set_lpjmlstats_settings}
\title{Set the package settings for lpjmlstats}
\usage{
set_lpjmlstats_settings(...)
}
\arguments{
\item{...}{Variable arguments to specify settings. The function accepts
the following options:
\itemize{
\item \code{year_subset}: A vector of years passed to the
subset argument of the \link{read_io} function. Defaults to reading
all years. Specifying a subset of years can improve performance.

\item \code{graphics_device}:  A character string specifying
the graphics device to be used for plotting the benchmarking results.
Defaults to "png". Use "pdf" for vector graphics.

\item \code{pdf_plot_dpi}:  A numeric value specifying the
DPI for the PDF document.

\item \code{unit_table_path}: A character string specifying
the path to the unit conversion table (a .csv file). Defaults to the
conversion table in the package's \code{inst} folder. The specified
file must exist and be a .csv file.
}}
}
\description{
This function configures various settings for the lpjmlstats package.
}
\examples{
\dontrun{
set_lpjmlstats_settings(year_subset = 1:5, graphics_device = "pdf")
set_lpjmlstats_settings(unit_table_path = "path/to/my_table.csv", pdf_plot_dpi = 300)
}

}
