% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLDataCalc.R
\name{subset_calc}
\alias{subset_calc}
\title{Subset an LPJmLDataCalc object}
\usage{
subset_calc(x, ...)
}
\arguments{
\item{x}{LPJmLDataCalc object.}

\item{...}{Parameters that are passed to \link[lpjmlkit]{subset.LPJmLData}.}
}
\value{
An \code{\link{LPJmLDataCalc}} object.
}
\description{
Function to subset an LPJmLDataCalc object. The function acts as a wrapper
of \link[lpjmlkit]{subset.LPJmLData} from lpjmlkit, but outputs an
\code{\link{LPJmLDataCalc}} object, in particular keeping its unit.
}
