% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metric.R
\name{TimeAvg}
\alias{TimeAvg}
\title{TimeAvg}
\description{
TimeAvg metric.
}
\section{Super class}{
\code{\link[lpjmlstats:Metric]{lpjmlstats::Metric}} -> \code{TimeAvg}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimeAvg-summarize}{\code{TimeAvg$summarize()}}
\item \href{#method-TimeAvg-compare}{\code{TimeAvg$compare()}}
\item \href{#method-TimeAvg-clone}{\code{TimeAvg$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="add_comparison"><a href='../../lpjmlstats/html/Metric.html#method-Metric-add_comparison'><code>lpjmlstats::Metric$add_comparison()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="capture_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-capture_summary'><code>lpjmlstats::Metric$capture_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="store_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-store_summary'><code>lpjmlstats::Metric$store_summary()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeAvg-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-TimeAvg-summarize}{}}}
\subsection{Method \code{summarize()}}{
Average all timesteps of the simulation period.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeAvg$summarize(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Raw data to be summarized}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeAvg-compare"></a>}}
\if{latex}{\out{\hypertarget{method-TimeAvg-compare}{}}}
\subsection{Method \code{compare()}}{
Compare the baseline summary with the under test summaries by
subtracting the baseline from the under test.
Delete the baseline summaries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeAvg$compare(eval_grp)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{eval_grp}}{Evaluation group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeAvg-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeAvg-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeAvg$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
