% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLDataCalc_aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate an LPJmLDataCalc object}
\usage{
aggregate(x, ref_area = "terr_area", ...)
}
\arguments{
\item{x}{\code{\link{LPJmLDataCalc}} object to be aggregated.}

\item{ref_area}{Specifies the reference area to be used as
a multiplier for the \code{weighted_sum} and \code{weighted_mean} aggregation methods.
If the data is given as an area density (i.e. per m2) the reference area
should be the area of each cell on which this area density "lives", assuming
it has the given value only on that area and the value zero elsewhere
(mathematically this is the support of the area density).}

\item{...}{One or several key-value combinations where keys represent the
dimension names and values specify the target aggregation units
and optionally also the summary statistic to be used. If the value
is a string, it is interpreted as the target aggregation unit and the method
defaults to \code{weighted_sum} and \code{mean} for the cell and time dimension,
respectively. If the value is itself a list, it must contain the
keys \code{to} and \code{stat} specifying the aggregation unit and method.
The aggregation units for the cell dimension can be either a string with
the following options
\itemize{
\item \code{countries}: The regions defined in the countries of the world file
\item \code{global}: A dynamically created region that fully contains all cells
of the grid
or an \code{\link{LPJmLRegionData}} object specifying different regions as columns of
its region matrix.
For the time dimension the only available aggregation unit is \code{full} which
aggregates the data over the full simulation period.
The aggregation method for space has the following options:
\item \code{sum}: The values of all cells belonging to each region are summed up.
If a cell belongs to a region only partially, we assume
that the quantity is distributed uniformly over the cell area and
multiply the value by the fraction of the cell that is part of the region
before summing up.
\item \code{mean}: First sums up the values of all cells belonging to each region
as described for \code{sum} and then divides by the number of cells belonging to
the region. Again we account for partial belonging of cells to regions
(if it exists) by only counting the fraction of the cell that is part of
the region in the divisor.
\item \code{weighted_sum}: Similar to the \code{sum} option but multiplies the value of
each cell by a reference area before summing up. The reference area
default is the \code{terr_area} output which needs to exist in the same directory
as the output to be aggregated. Other reference areas can be specified
by setting the \code{reference_area} parameter.
\item \code{weighted_mean}: Similar to the \code{mean} option but multiplies the value of
each cell by a reference area before summing up. Also,
the resulting sum is then divided by the total reference area of each
region instead of the number of cells.
}}
}
\value{
An aggregated \code{\link{LPJmLDataCalc}} object.
}
\description{
Function to aggregate the full data of an \code{\link{LPJmLDataCalc}} object by
applying summary statistics along the cell and/or time dimensions.
}
