% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLDataCalc.R
\name{.as_LPJmLDataCalc}
\alias{.as_LPJmLDataCalc}
\title{Coerce an LPJmLData object into an LPJmLDataCalc object}
\usage{
.as_LPJmLDataCalc(obj)
}
\arguments{
\item{obj}{LPJmLData object.
For internal package development use the obj can also be an array with the dimension
\enumerate{
\item cell/region, 2. years, 3. bands. The items of the time dimension are then assumed to be different years.
}}
}
\value{
An LPJmLDataCalc object.
}
\description{
Function to coerce (convert) an \code{\link{LPJmLData}} object into an
LPJmLDataCalc object with extended functionality.
}
