% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metric_subclasses.R
\name{GlobSumTimeseries}
\alias{GlobSumTimeseries}
\title{GlobSumTimeseries}
\description{
GlobSumTimeseries metric.
See \link{Metric} for the documentation of metrics in general.
}
\section{Super class}{
\code{\link[lpjmlstats:Metric]{lpjmlstats::Metric}} -> \code{GlobSumTimeseries}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{m_options}}{List of metric options specific to this metric
\itemize{
\item \code{font_size} integer, font size of the table
\item \code{name_trunc} integer, indicating when to truncate the band names
band names
\item \code{year_subset}: character vector, defines which calander years the metric considers,
i.e. a data subset that the metric works with; e.g. c("1995", "1996").
}}

\item{\code{title}}{Section header used in the report}

\item{\code{description}}{Description used in the report}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GlobSumTimeseries-summarize}{\code{GlobSumTimeseries$summarize()}}
\item \href{#method-GlobSumTimeseries-plot}{\code{GlobSumTimeseries$plot()}}
\item \href{#method-GlobSumTimeseries-arrange_plots}{\code{GlobSumTimeseries$arrange_plots()}}
\item \href{#method-GlobSumTimeseries-clone}{\code{GlobSumTimeseries$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="add_compare_meta"><a href='../../lpjmlstats/html/Metric.html#method-Metric-add_compare_meta'><code>lpjmlstats::Metric$add_compare_meta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="add_comparisons"><a href='../../lpjmlstats/html/Metric.html#method-Metric-add_comparisons'><code>lpjmlstats::Metric$add_comparisons()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="capture_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-capture_summary'><code>lpjmlstats::Metric$capture_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="compare"><a href='../../lpjmlstats/html/Metric.html#method-Metric-compare'><code>lpjmlstats::Metric$compare()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="generate_report_content"><a href='../../lpjmlstats/html/Metric.html#method-Metric-generate_report_content'><code>lpjmlstats::Metric$generate_report_content()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="print_metric_description"><a href='../../lpjmlstats/html/Metric.html#method-Metric-print_metric_description'><code>lpjmlstats::Metric$print_metric_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="print_metric_header"><a href='../../lpjmlstats/html/Metric.html#method-Metric-print_metric_header'><code>lpjmlstats::Metric$print_metric_header()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="print_year_subset"><a href='../../lpjmlstats/html/Metric.html#method-Metric-print_year_subset'><code>lpjmlstats::Metric$print_year_subset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="store_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-store_summary'><code>lpjmlstats::Metric$store_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="transform_lpjml_calcs"><a href='../../lpjmlstats/html/Metric.html#method-Metric-transform_lpjml_calcs'><code>lpjmlstats::Metric$transform_lpjml_calcs()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobSumTimeseries-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-GlobSumTimeseries-summarize}{}}}
\subsection{Method \code{summarize()}}{
Take a global weighted sum of the output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobSumTimeseries$summarize(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{LPJmLDataCalc object to be summarized}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A summarized \link{LPJmLDataCalc} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobSumTimeseries-plot"></a>}}
\if{latex}{\out{\hypertarget{method-GlobSumTimeseries-plot}{}}}
\subsection{Method \code{plot()}}{
Create a time series plot of the results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobSumTimeseries$plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of time series ggplots
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobSumTimeseries-arrange_plots"></a>}}
\if{latex}{\out{\hypertarget{method-GlobSumTimeseries-arrange_plots}{}}}
\subsection{Method \code{arrange_plots()}}{
Arrange the time series plots side by side
with legends pooled together in the top left
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobSumTimeseries$arrange_plots(plotlist)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plotlist}}{List of time series ggplots}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobSumTimeseries-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GlobSumTimeseries-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobSumTimeseries$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
