% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLDataCalc.R
\name{LPJmLDataCalc}
\alias{LPJmLDataCalc}
\title{LPJmLDataCalc}
\description{
An extended LPJmLData class that enables arithmetic and statistics.
}
\section{Super class}{
\code{\link[lpjmlkit:LPJmLData]{lpjmlkit::LPJmLData}} -> \code{LPJmLDataCalc}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{the data array}

\item{\code{.data_with_unit}}{Returns the internal enclosed unit object
!Internal method only to be used for package development!}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LPJmLDataCalc-new}{\code{LPJmLDataCalc$new()}}
\item \href{#method-LPJmLDataCalc-aggregate}{\code{LPJmLDataCalc$aggregate()}}
\item \href{#method-LPJmLDataCalc-get_ref_area}{\code{LPJmLDataCalc$get_ref_area()}}
\item \href{#method-LPJmLDataCalc-plot}{\code{LPJmLDataCalc$plot()}}
\item \href{#method-LPJmLDataCalc-.check_internal_integrity}{\code{LPJmLDataCalc$.check_internal_integrity()}}
\item \href{#method-LPJmLDataCalc-.plot_aggregated}{\code{LPJmLDataCalc$.plot_aggregated()}}
\item \href{#method-LPJmLDataCalc-.add}{\code{LPJmLDataCalc$.add()}}
\item \href{#method-LPJmLDataCalc-.subtract}{\code{LPJmLDataCalc$.subtract()}}
\item \href{#method-LPJmLDataCalc-.multiply}{\code{LPJmLDataCalc$.multiply()}}
\item \href{#method-LPJmLDataCalc-.divide}{\code{LPJmLDataCalc$.divide()}}
\item \href{#method-LPJmLDataCalc-.convert_unit}{\code{LPJmLDataCalc$.convert_unit()}}
\item \href{#method-LPJmLDataCalc-.set_unit}{\code{LPJmLDataCalc$.set_unit()}}
\item \href{#method-LPJmLDataCalc-apply_unit_conversion_table}{\code{LPJmLDataCalc$apply_unit_conversion_table()}}
\item \href{#method-LPJmLDataCalc-set_sim_identifier}{\code{LPJmLDataCalc$set_sim_identifier()}}
\item \href{#method-LPJmLDataCalc-get_sim_identifier}{\code{LPJmLDataCalc$get_sim_identifier()}}
\item \href{#method-LPJmLDataCalc-add_grid}{\code{LPJmLDataCalc$add_grid()}}
\item \href{#method-LPJmLDataCalc-clone}{\code{LPJmLDataCalc$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id=".__set_data__"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-.__set_data__'><code>lpjmlkit::LPJmLData$.__set_data__()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id=".__set_grid__"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-.__set_grid__'><code>lpjmlkit::LPJmLData$.__set_grid__()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="as_array"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-as_array'><code>lpjmlkit::LPJmLData$as_array()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="as_raster"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-as_raster'><code>lpjmlkit::LPJmLData$as_raster()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="as_terra"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-as_terra'><code>lpjmlkit::LPJmLData$as_terra()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="as_tibble"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-as_tibble'><code>lpjmlkit::LPJmLData$as_tibble()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="dim"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-dim'><code>lpjmlkit::LPJmLData$dim()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="dimnames"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-dimnames'><code>lpjmlkit::LPJmLData$dimnames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="length"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-length'><code>lpjmlkit::LPJmLData$length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="print"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-print'><code>lpjmlkit::LPJmLData$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="subset"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-subset'><code>lpjmlkit::LPJmLData$subset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="summary"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-summary'><code>lpjmlkit::LPJmLData$summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLData" data-id="transform"><a href='../../lpjmlkit/html/LPJmLData.html#method-LPJmLData-transform'><code>lpjmlkit::LPJmLData$transform()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-new"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-new}{}}}
\subsection{Method \code{new()}}{
Create a new LPJmLDataCalc object; to be used internally or explicitly
!Internal method only to be used for package development!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$new(lpjml_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lpjml_data}}{an LPJmLData object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-aggregate"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-aggregate}{}}}
\subsection{Method \code{aggregate()}}{
See \link[lpjmlstats]{aggregate}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$aggregate(ref_area = "terr_area", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ref_area}}{See \link[lpjmlstats]{aggregate}.}

\item{\code{...}}{See \link[lpjmlstats]{aggregate}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-get_ref_area"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-get_ref_area}{}}}
\subsection{Method \code{get_ref_area()}}{
Get the reference area of the LPJmLDataCalc object.
For an area density variable the reference area should
be the area of each cell on which the variable is defined.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$get_ref_area(ref_area)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ref_area}}{A string that can be
\itemize{
\item \code{terr_area} terrestrial area (land area including inland water bodies)
\item \code{cell_area} full area of each cell
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{\link{LPJmLDataCalc}} object with the reference area as variable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-plot"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-plot}{}}}
\subsection{Method \code{plot()}}{
Plot an LPJmLDataCalc object

The function acts a wrapper of \link[lpjmlkit]{plot.LPJmLData} from
lpjmlkit, but allows for plotting data in more formats.

In case of non-aggregated data
\link[lpjmlkit]{plot.LPJmLData}
is directly called. In case of aggregated data the
value for each region is assigned to all pixels that belong
to the region.
If a pixel belong to a region only partially, the value is
multiplied by the
fraction of that pixel belonging to the region. If a pixel belongs to
multiple regions, the sum of all respective region values
(multiplied by the fractions) is taken.
The pixel values are then again plotted with
\link[lpjmlkit]{plot.LPJmLData}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link[lpjmlkit]{LPJmLData} plot method.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-.check_internal_integrity"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-.check_internal_integrity}{}}}
\subsection{Method \code{.check_internal_integrity()}}{
Check consistency of data and meta data
!Internal method only to be used for package development!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$.check_internal_integrity()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-.plot_aggregated"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-.plot_aggregated}{}}}
\subsection{Method \code{.plot_aggregated()}}{
Plot aggregated data.
Performs a very simple disaggregation to create LPJmLData obj
that can be plotted with plot.LPJmLData.
For each pixel the values of all regions that contain
the pixel are multiplied by the fractions and summed up.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$.plot_aggregated(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments to be passed to plot.LPJmLData}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-.add"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-.add}{}}}
\subsection{Method \code{.add()}}{
Addition of two LPJmLDataCalc objects
!Internal method only to be used for package development!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$.add(lpjml_calc_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lpjml_calc_obj}}{An \code{LPJmLData} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-.subtract"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-.subtract}{}}}
\subsection{Method \code{.subtract()}}{
Subtraction of two LPJmLDataCalc objects
!Internal method only to be used for package development!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$.subtract(lpjml_calc_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lpjml_calc_obj}}{An \code{LPJmLData} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-.multiply"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-.multiply}{}}}
\subsection{Method \code{.multiply()}}{
Multiplication of two LPJmLDataCalc objects
!Internal method only to be used for package development!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$.multiply(lpjml_calc_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lpjml_calc_obj}}{An \code{LPJmLData} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-.divide"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-.divide}{}}}
\subsection{Method \code{.divide()}}{
Division of two LPJmLDataCalc objects
!Internal method only to be used for package development!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$.divide(lpjml_calc_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lpjml_calc_obj}}{An \code{LPJmLData} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-.convert_unit"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-.convert_unit}{}}}
\subsection{Method \code{.convert_unit()}}{
Unit conversion of LPJmLDataCalc object
!Internal method only to be used for package development!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$.convert_unit(unit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unit}}{A string with the unit to convert to.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-.set_unit"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-.set_unit}{}}}
\subsection{Method \code{.set_unit()}}{
Set unit of LPJmLDataCalc object
!Internal method only to be used for package development!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$.set_unit(unit_str)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unit_str}}{A string with the unit to be set.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-apply_unit_conversion_table"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-apply_unit_conversion_table}{}}}
\subsection{Method \code{apply_unit_conversion_table()}}{
Apply unit conversion from conversion table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$apply_unit_conversion_table(path_to_table = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path_to_table}}{A string with the path to the conversion table.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-set_sim_identifier"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-set_sim_identifier}{}}}
\subsection{Method \code{set_sim_identifier()}}{
!Internal method only to be used for package development!
Set the simulation identifier to the LPJmLDataCalc object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$set_sim_identifier(identifier)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{identifier}}{A string with the identifier to set.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-get_sim_identifier"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-get_sim_identifier}{}}}
\subsection{Method \code{get_sim_identifier()}}{
!Internal method only to be used for package development!
Get the simulation identifier of the LPJmLDataCalc object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$get_sim_identifier()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string with the simulation identifier.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-add_grid"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-add_grid}{}}}
\subsection{Method \code{add_grid()}}{
Add a grid to the LPJmLDataCalc object
Wrapper for the \code{add_grid} method of the \code{LPJmLData} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$add_grid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLDataCalc-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLDataCalc-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLDataCalc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
