% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLMetaDataCalc.R
\name{LPJmLMetaDataCalc}
\alias{LPJmLMetaDataCalc}
\title{LPJmL meta data class}
\description{
A meta data container for the LPJmLDataCalc class that extends the
\link[lpjmlkit]{LPJmLMetaData} such that aggregation can be tracked.
}
\section{Super class}{
\code{\link[lpjmlkit:LPJmLMetaData]{lpjmlkit::LPJmLMetaData}} -> \code{LPJmLMetaDataCalc}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{space_aggregation}}{Indication weather the data has been
subject to space aggregation.}

\item{\code{time_aggregation}}{Indication weather the data has been
subject to time aggregation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LPJmLMetaDataCalc-new}{\code{LPJmLMetaDataCalc$new()}}
\item \href{#method-LPJmLMetaDataCalc-.__set_space_aggregation__}{\code{LPJmLMetaDataCalc$.__set_space_aggregation__()}}
\item \href{#method-LPJmLMetaDataCalc-.__set_time_aggregation__}{\code{LPJmLMetaDataCalc$.__set_time_aggregation__()}}
\item \href{#method-LPJmLMetaDataCalc-print}{\code{LPJmLMetaDataCalc$print()}}
\item \href{#method-LPJmLMetaDataCalc-clone}{\code{LPJmLMetaDataCalc$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLMetaData" data-id=".__init_grid__"><a href='../../lpjmlkit/html/LPJmLMetaData.html#method-LPJmLMetaData-.__init_grid__'><code>lpjmlkit::LPJmLMetaData$.__init_grid__()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLMetaData" data-id=".__set_attribute__"><a href='../../lpjmlkit/html/LPJmLMetaData.html#method-LPJmLMetaData-.__set_attribute__'><code>lpjmlkit::LPJmLMetaData$.__set_attribute__()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLMetaData" data-id=".__transform_space_format__"><a href='../../lpjmlkit/html/LPJmLMetaData.html#method-LPJmLMetaData-.__transform_space_format__'><code>lpjmlkit::LPJmLMetaData$.__transform_space_format__()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLMetaData" data-id=".__transform_time_format__"><a href='../../lpjmlkit/html/LPJmLMetaData.html#method-LPJmLMetaData-.__transform_time_format__'><code>lpjmlkit::LPJmLMetaData$.__transform_time_format__()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLMetaData" data-id=".__update_subset__"><a href='../../lpjmlkit/html/LPJmLMetaData.html#method-LPJmLMetaData-.__update_subset__'><code>lpjmlkit::LPJmLMetaData$.__update_subset__()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLMetaData" data-id="as_header"><a href='../../lpjmlkit/html/LPJmLMetaData.html#method-LPJmLMetaData-as_header'><code>lpjmlkit::LPJmLMetaData$as_header()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlkit" data-topic="LPJmLMetaData" data-id="as_list"><a href='../../lpjmlkit/html/LPJmLMetaData.html#method-LPJmLMetaData-as_list'><code>lpjmlkit::LPJmLMetaData$as_list()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLMetaDataCalc-new"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLMetaDataCalc-new}{}}}
\subsection{Method \code{new()}}{
Initialize the LPJmLMetaDataCalc object by copying all
private attributes from an LPJmLMetaData object to private attributes
of this object.
!Internal method only to be used for package development!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLMetaDataCalc$new(lpjml_meta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lpjml_meta}}{an LPJmLMetaData object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLMetaDataCalc-.__set_space_aggregation__"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLMetaDataCalc-.__set_space_aggregation__}{}}}
\subsection{Method \code{.__set_space_aggregation__()}}{
Save in metadata that data is in space_aggregation format
!Internal method only to be used for package development!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLMetaDataCalc$.__set_space_aggregation__(agg_method)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{agg_method}}{string indicating the aggregation method}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLMetaDataCalc-.__set_time_aggregation__"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLMetaDataCalc-.__set_time_aggregation__}{}}}
\subsection{Method \code{.__set_time_aggregation__()}}{
Save in metadata that data is in time_aggregation format
!Internal method only to be used for package development!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLMetaDataCalc$.__set_time_aggregation__(agg_method)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{agg_method}}{string indicating the aggregation method}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLMetaDataCalc-print"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLMetaDataCalc-print}{}}}
\subsection{Method \code{print()}}{
Wrapper for \code{\link{LPJmLMetaData}} print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLMetaDataCalc$print(spaces = "", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{spaces}}{string of spaces to be printed as prefix}

\item{\code{...}}{additional arguments passed to \code{\link{LPJmLMetaData}} print method}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPJmLMetaDataCalc-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LPJmLMetaDataCalc-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLMetaDataCalc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
