% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpjmlstats-package.R
\docType{package}
\name{lpjmlstats-package}
\alias{lpjmlstats}
\alias{lpjmlstats-package}
\title{lpjmlstats: Statistical tools for LPJmL data analysis}
\description{
This package provides statistical tools for LPJmL data analysis to be used for benchmarking LPJmL outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PIK-LPJmL/lpjmlstats}
  \item Report bugs at \url{https://github.com/PIK-LPJmL/lpjmlstats/issues}
}

}
\author{
\strong{Maintainer}: David Hötten \email{davidho@pik-potsdam.de}

Authors:
\itemize{
  \item Jannes Breier \email{jannes.breier@pik-potsdam.de} (\href{https://orcid.org/0000-0002-9055-6904}{ORCID})
  \item Christoph Müller \email{cmueller@pik-potsdam.de} (\href{https://orcid.org/0000-0002-9491-3550}{ORCID})
  \item Stephen B. Wirth \email{stephen.wirth@pik-potsdam.de} (\href{https://orcid.org/0000-0003-3090-3318}{ORCID})
}

}
\keyword{internal}
