% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetricVarGrp.R
\name{Metric}
\alias{Metric}
\title{Metric super class}
\description{
A metric is
a) a structured, generic processing pipeline to calculate
numerical indicators for sets of lpjml outputs as well as
to display these indicators in a report, and
b) a structured data container that stores the numerical indicators resulting
from a).

\strong{a)} A metric defines a procedure to
\enumerate{
\item \strong{summarize} a complex, multidimensional LPJmL
output in a meaningfull, potentially opinionated way,
typically involving the reduction of its time and or space dimension,
\item \strong{compare} how this summary statistic of an output variable changes,
when going from an established baseline LPJmL version or configuration to
a new version or configuration currently under test,
\item \strong{plot} the results of 1. and 2. as a figure or table and
\item \strong{arrange} these plots in a report,
(e.g. styling, side by side arrangement)
}

See \link{benchmark} on how these steps come into play in the
benchmarking process.

\strong{b)}  The summarized outputs as well as comparisons that are stored
a metric are grouped by the different lpjml variables. A so called variable
group (var_grp) contains
\enumerate{
\item the summary of the baseline output of that variable
\item a list of all summarized under test outputs of that variable
\item a list of compare items (e.g. difference, relative difference). Each
compare item is a list of comparisons of the baseline summary with
each under test summary using the specific method of that item.
}

All variable groups are stored in the var_grp_list attribute of the metric.

As the cornerstone of the benchmarking process, all metrics illuminate
the change of LPJmL results from different angles,
and should together provide a comprehensive picture of the effects of
modifications in code or settings.

See \link{GlobSumTimeAvgTable} for a typical example of a metric.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{var_grp_list}}{List of variable groups. Each variable group contains the summaries
and the comparisons for one variable.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Metric-summarize}{\code{Metric$summarize()}}
\item \href{#method-Metric-compare}{\code{Metric$compare()}}
\item \href{#method-Metric-plot}{\code{Metric$plot()}}
\item \href{#method-Metric-arrange_plots}{\code{Metric$arrange_plots()}}
\item \href{#method-Metric-capture_summary}{\code{Metric$capture_summary()}}
\item \href{#method-Metric-store_summary}{\code{Metric$store_summary()}}
\item \href{#method-Metric-add_comparison}{\code{Metric$add_comparison()}}
\item \href{#method-Metric-apply_to_all_lpjml_calcs}{\code{Metric$apply_to_all_lpjml_calcs()}}
\item \href{#method-Metric-generate_report_content}{\code{Metric$generate_report_content()}}
\item \href{#method-Metric-print_metric_header}{\code{Metric$print_metric_header()}}
\item \href{#method-Metric-print_metric_description}{\code{Metric$print_metric_description()}}
\item \href{#method-Metric-clone}{\code{Metric$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-summarize}{}}}
\subsection{Method \code{summarize()}}{
Pipeline to summarize the raw data.
Will be overwritten by the individual metric subclasses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$summarize(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Raw data to be summarized}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-compare"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-compare}{}}}
\subsection{Method \code{compare()}}{
Pipeline to compare the baseline summary with each under test summary
stored in the metric.
Will be overwritten by the individual metric subclasses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$compare(var_grp)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var_grp}}{variable group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-plot}{}}}
\subsection{Method \code{plot()}}{
Function to plot the results of the metric.
Will be overwritten by the individual metric subclasses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$plot(var_grp)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var_grp}}{variable group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-arrange_plots"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-arrange_plots}{}}}
\subsection{Method \code{arrange_plots()}}{
Function to arrgange all plots of the metric in the respective
section of the report.
Will be overwritten by the individual metric subclasses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$arrange_plots(var_grp)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var_grp}}{variable group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-capture_summary"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-capture_summary}{}}}
\subsection{Method \code{capture_summary()}}{
!Package internal method!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$capture_summary(data, var, type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Raw data to be summarized}

\item{\code{var}}{Variable name}

\item{\code{type}}{Type of data ("baseline" or "under_test")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-store_summary"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-store_summary}{}}}
\subsection{Method \code{store_summary()}}{
!Package internal method!
Store the summary in the variable group
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$store_summary(summary, var, type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summary}}{Summary to be stored}

\item{\code{var}}{Variable name}

\item{\code{type}}{Type of data ("baseline" or "under_test")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-add_comparison"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-add_comparison}{}}}
\subsection{Method \code{add_comparison()}}{
!Package internal method!
Compare and store the comparison in the variable group
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$add_comparison()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{Variable name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-apply_to_all_lpjml_calcs"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-apply_to_all_lpjml_calcs}{}}}
\subsection{Method \code{apply_to_all_lpjml_calcs()}}{
!Package internal method!
Apply function to all lpjml_calcs in all eval groups and lists
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$apply_to_all_lpjml_calcs(fun, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{Function to apply}

\item{\code{...}}{Additional arguments passed to fun}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-generate_report_content"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-generate_report_content}{}}}
\subsection{Method \code{generate_report_content()}}{
!Package internal method!
Generate the full report content of the metric.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$generate_report_content()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-print_metric_header"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-print_metric_header}{}}}
\subsection{Method \code{print_metric_header()}}{
!Package internal method!
Function to print the metric header.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$print_metric_header()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-print_metric_description"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-print_metric_description}{}}}
\subsection{Method \code{print_metric_description()}}{
!Package internal method!
Function to print the metric description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$print_metric_description()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
