% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_settings.R
\name{set_lpjmlstats_settings}
\alias{set_lpjmlstats_settings}
\title{Set the package settings for lpjmlstats}
\usage{
set_lpjmlstats_settings(
  year_subset = NULL,
  graphics_device = NULL,
  pdf_plot_dpi = NULL,
  unit_table_path = NULL
)
}
\arguments{
\item{year_subset}{vector of years passed to the subset argument of the
\link{read_io} function to read in the output files for the
benchmarking. The default is to read in all years.
Note that the benchmarking will run much faster if
only a subset of years needs to be read in and processed.}

\item{graphics_device}{character string, the graphics device to be used for
plotting the benchmarking results. The default is "png".
Using "pdf" will result in plots using vector graphics, which
allow infinite zooming without loss of quality.}

\item{pdf_plot_dpi}{numeric, the dpi that is used for the graphics device.}

\item{unit_table_path}{character string, the path to the unit
conversion table. The table must be a .csv file. The default is to use
the conversion table in the inst folder of the package.}
}
\description{
Set the package settings for lpjmlstats
}
