% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metric_subclasses.R
\name{TimeAvgMap}
\alias{TimeAvgMap}
\title{TimeAvgMap}
\description{
TimeAvgMap metric
}
\section{Super class}{
\code{\link[lpjmlstats:Metric]{lpjmlstats::Metric}} -> \code{TimeAvgMap}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{m_options}}{List of options for the map plot
\itemize{
\item \code{m_options$font_size} integer, font size of the map plot
\item \code{m_options$name_trunc} integer, indicating when to truncate the
band names
\item \code{m_options$highlight} vector of strings, indicating which variables
should receive a larger full width plot
}}

\item{\code{title}}{Section header used in the report}

\item{\code{description}}{Description used in the report}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimeAvgMap-summarize}{\code{TimeAvgMap$summarize()}}
\item \href{#method-TimeAvgMap-compare}{\code{TimeAvgMap$compare()}}
\item \href{#method-TimeAvgMap-plot}{\code{TimeAvgMap$plot()}}
\item \href{#method-TimeAvgMap-arrange_plots}{\code{TimeAvgMap$arrange_plots()}}
\item \href{#method-TimeAvgMap-clone}{\code{TimeAvgMap$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="add_comparison"><a href='../../lpjmlstats/html/Metric.html#method-Metric-add_comparison'><code>lpjmlstats::Metric$add_comparison()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="apply_to_all_lpjml_calcs"><a href='../../lpjmlstats/html/Metric.html#method-Metric-apply_to_all_lpjml_calcs'><code>lpjmlstats::Metric$apply_to_all_lpjml_calcs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="capture_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-capture_summary'><code>lpjmlstats::Metric$capture_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="generate_report_content"><a href='../../lpjmlstats/html/Metric.html#method-Metric-generate_report_content'><code>lpjmlstats::Metric$generate_report_content()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="print_metric_description"><a href='../../lpjmlstats/html/Metric.html#method-Metric-print_metric_description'><code>lpjmlstats::Metric$print_metric_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="print_metric_header"><a href='../../lpjmlstats/html/Metric.html#method-Metric-print_metric_header'><code>lpjmlstats::Metric$print_metric_header()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="store_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-store_summary'><code>lpjmlstats::Metric$store_summary()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeAvgMap-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-TimeAvgMap-summarize}{}}}
\subsection{Method \code{summarize()}}{
Take the mean over the simulation period for each cell.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeAvgMap$summarize(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{LpjmlDataCalc object to be summarized}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A summarized \link{LPJmLDataCalc} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeAvgMap-compare"></a>}}
\if{latex}{\out{\hypertarget{method-TimeAvgMap-compare}{}}}
\subsection{Method \code{compare()}}{
Compare the baseline summary with the under test summaries by
subtracting the baseline from the under test.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeAvgMap$compare(var_grp)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var_grp}}{variable group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeAvgMap-plot"></a>}}
\if{latex}{\out{\hypertarget{method-TimeAvgMap-plot}{}}}
\subsection{Method \code{plot()}}{
Create a map plot with country border overlay.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeAvgMap$plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of map ggplots
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeAvgMap-arrange_plots"></a>}}
\if{latex}{\out{\hypertarget{method-TimeAvgMap-arrange_plots}{}}}
\subsection{Method \code{arrange_plots()}}{
Arrange the map plots side by side
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeAvgMap$arrange_plots(plotlist)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plotlist}}{List of map ggplots}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeAvgMap-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeAvgMap-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeAvgMap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
