% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metric_subclasses.R
\name{TimeAvgMapWithAbsUndertest}
\alias{TimeAvgMapWithAbsUndertest}
\title{TimeAvgMapWithAbsUndertest}
\description{
TimeAvgMapWithAbsUndertest metric.
See \link{Metric} for the documentation of metrics in general.
}
\section{Super classes}{
\code{\link[lpjmlstats:Metric]{lpjmlstats::Metric}} -> \code{\link[lpjmlstats:TimeAvgMap]{lpjmlstats::TimeAvgMap}} -> \code{\link[lpjmlstats:TimeAvgMapWithAbs]{lpjmlstats::TimeAvgMapWithAbs}} -> \code{TimeAvgMapWithAbsUndertest}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{m_options}}{List of metric options specific to this metric:
\itemize{
\item \code{font_size}: integer, font size of the map plot (default 6)
\item \code{highlight}: vector of strings, indicating which variables
should be highlighted in the report, that is receive a larger plot at the
beginning of report content of the metric.
All variables with a name that contains at least one these strings as a substring,
will not be plotted in the plotgrid (see \code{num_cols}) but before the plot grid starts.
These plots are allowed to extent to full page width. (default NULL)
\item \code{scale_fill_quantiles}: quantiles used to determine the lower and upper
limits for the values in the map plot (default c(0.05, 0.95))
\item \code{sep_cmp_lims}: logical, if TRUE not all plots of a var_grp
will have the same limits anymore, but the compare plots
have their own separate limits (default TRUE)
\item \code{scale_fill_n_breaks}: number of breaks for each arm of the diverging
color scale (default 5)
\item \code{scale_fill_tick_label_angle}: integer, angle in degrees for the fill axis (colorbar)
tick labels (default 45)
\item \code{year_subset}: character vector, defines which calendar years the metric considers,
i.e., a data subset that the metric works with; e.g., c("1995", "1996") (default 1991:2000).
\item \code{cell_subset}: character vector, defines which cells to subset (default NULL)
\item \code{num_cols}: integer, number of columns in the plot grid in the report (default 2)
\item \code{var_subheading}: logical, if TRUE, a linebreak and a subheading will
be inserted before plots for a new variable are added to the report.
with the name of the variable will be added. Both things are intended to visually seperate
the plots of different variables and to better organize the report,
especially if the metric generates many plots for each variable. (default FALSE)
\item \code{band_subheading}: analogous to var_subheading but for bands (default FALSE)
}}

\item{\code{title}}{Section header used in the report}

\item{\code{description}}{Description used in the report}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimeAvgMapWithAbsUndertest-compare}{\code{TimeAvgMapWithAbsUndertest$compare()}}
\item \href{#method-TimeAvgMapWithAbsUndertest-clone}{\code{TimeAvgMapWithAbsUndertest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="add_compare_meta"><a href='../../lpjmlstats/html/Metric.html#method-Metric-add_compare_meta'><code>lpjmlstats::Metric$add_compare_meta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="add_comparisons"><a href='../../lpjmlstats/html/Metric.html#method-Metric-add_comparisons'><code>lpjmlstats::Metric$add_comparisons()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="capture_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-capture_summary'><code>lpjmlstats::Metric$capture_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="generate_report_content"><a href='../../lpjmlstats/html/Metric.html#method-Metric-generate_report_content'><code>lpjmlstats::Metric$generate_report_content()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="print_metric_description"><a href='../../lpjmlstats/html/Metric.html#method-Metric-print_metric_description'><code>lpjmlstats::Metric$print_metric_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="print_metric_header"><a href='../../lpjmlstats/html/Metric.html#method-Metric-print_metric_header'><code>lpjmlstats::Metric$print_metric_header()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="print_year_subset"><a href='../../lpjmlstats/html/Metric.html#method-Metric-print_year_subset'><code>lpjmlstats::Metric$print_year_subset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="store_summary"><a href='../../lpjmlstats/html/Metric.html#method-Metric-store_summary'><code>lpjmlstats::Metric$store_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="Metric" data-id="transform_lpjml_calcs"><a href='../../lpjmlstats/html/Metric.html#method-Metric-transform_lpjml_calcs'><code>lpjmlstats::Metric$transform_lpjml_calcs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="TimeAvgMap" data-id="summarize"><a href='../../lpjmlstats/html/TimeAvgMap.html#method-TimeAvgMap-summarize'><code>lpjmlstats::TimeAvgMap$summarize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="TimeAvgMapWithAbs" data-id="arrange_plots"><a href='../../lpjmlstats/html/TimeAvgMapWithAbs.html#method-TimeAvgMapWithAbs-arrange_plots'><code>lpjmlstats::TimeAvgMapWithAbs$arrange_plots()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lpjmlstats" data-topic="TimeAvgMapWithAbs" data-id="plot"><a href='../../lpjmlstats/html/TimeAvgMapWithAbs.html#method-TimeAvgMapWithAbs-plot'><code>lpjmlstats::TimeAvgMapWithAbs$plot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeAvgMapWithAbsUndertest-compare"></a>}}
\if{latex}{\out{\hypertarget{method-TimeAvgMapWithAbsUndertest-compare}{}}}
\subsection{Method \code{compare()}}{
Compare the baseline summary with the under test summaries by
subtracting the baseline from the under test. The baseline is
then set to NULL so it is omitted when printing the plots.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeAvgMapWithAbsUndertest$compare(var_grp)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var_grp}}{variable group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeAvgMapWithAbsUndertest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeAvgMapWithAbsUndertest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeAvgMapWithAbsUndertest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
